+++ love/src/components/OLE/Exposure/ExposureAdd.jsx
@@ -103,7 +103,7 @@ export default class ExposureAdd extends Component {
   queryExposures() {
     const { selectedInstrument, selectedDayExposureStart, selectedDayExposureEnd, registryMap } = this.state;
     const startObsDay = Moment(selectedDayExposureStart).format(ISO_INTEGER_DATE_FORMAT);
-    const endObsDay = Moment(selectedDayExposureEnd).format(ISO_INTEGER_DATE_FORMAT);
+    const endObsDay = Moment(selectedDayExposureEnd).add(1, 'days').format(ISO_INTEGER_DATE_FORMAT);
     const registry = registryMap[selectedInstrument].split('_')[2];
 
     // Get the list of exposures