+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Version History
 ===============
 
+v5.24.9
+--------
+
+* Add 1 day to end obs day for querying exposures `<https://github.com/lsst-ts/LOVE-frontend/pull/510`_
+* Point LOVE weather station to read corresponding CSC `<https://github.com/lsst-ts/LOVE-frontend/pull/509>`_
+
 v5.24.8
 --------